+++ src/components/inputs/radioGroupInput.tsx
@@ -7,18 +7,20 @@ import { cn, getNestedProperty, setNestedProperty } from "@/lib/utils";
 import type { Settings } from "@/stores/settingsStore";
 import { defaultSettings, useSettingsStore } from "@/stores/settingsStore";
 import { RotateCcw } from "lucide-react";
-import type { ComponentProps } from "react";
+import type { ComponentProps, ReactNode } from "react";
 import { RadioGroup } from "../ui/radio-group";
 
 const RadioGroupInput = ({
   label,
   settingPath,
   selector,
+  description,
   ...props
 }: {
   label: string;
   settingPath?: keyof Settings | (string & {});
   selector?: (settings: Settings) => string;
+  description?: ReactNode;
 } & ComponentProps<typeof RadioGroup>) => {
   // Use selector if provided, otherwise use settingPath
   // When using settingPath, convert the value to string for the RadioGroup component
@@ -71,6 +73,9 @@ const RadioGroupInput = ({
 
         <RadioGroup value={value} {...props} />
       </div>
+      {description && (
+        <p className="text-muted-foreground/75 mt-1 text-sm">{description}</p>
+      )}
     </>
   );
 };
