+++ src/components/settings/general/hideBeatmapSetCoverWarning.tsx
@@ -1,4 +1,4 @@
-import { useSettingsStore } from "../../stores/settingsStore";
+import { useSettingsStore } from "@/stores/settingsStore";
 
 const HideBeatmapSetCoverWarning = () => {
   const hideBeatmapSetCovers = useSettingsStore.use.hideBeatmapSetCovers();
